# subScope: use a slice of tags instead of a map