--- a/tags.go
+++ b/tags.go
@@ -6,15 +6,201 @@ import (
 )
 
 type tagPair struct {
-	dimension string
-	value     string
+	key   string
+	value string
 }
 
 type tagSet []tagPair
 
 func (t tagSet) Len() int           { return len(t) }
 func (t tagSet) Swap(i, j int)      { t[i], t[j] = t[j], t[i] }
-func (t tagSet) Less(i, j int) bool { return t[i].dimension < t[j].dimension }
+func (t tagSet) Less(i, j int) bool { return t[i].key < t[j].key }
+
+func (t tagSet) CompareAndSwap(i, j int) {
+	if t[i].key > t[j].key {
+		t.Swap(i, j)
+	}
+}
+
+// Sort sorts the tagSet in place and is optimized for small (N <= 8) tagSets.
+func (t tagSet) Sort() {
+	// network sort generated with: https://pages.ripco.net/~jgamble/nw.html
+	switch len(t) {
+	case 0, 1:
+		return
+	case 2:
+		t.CompareAndSwap(0, 1)
+	case 3:
+		t.CompareAndSwap(1, 2)
+		t.CompareAndSwap(0, 2)
+		t.CompareAndSwap(0, 1)
+	case 4:
+		t.CompareAndSwap(0, 1)
+		t.CompareAndSwap(2, 3)
+		t.CompareAndSwap(0, 2)
+		t.CompareAndSwap(1, 3)
+		t.CompareAndSwap(1, 2)
+	case 5:
+		t.CompareAndSwap(0, 1)
+		t.CompareAndSwap(3, 4)
+		t.CompareAndSwap(2, 4)
+		t.CompareAndSwap(2, 3)
+		t.CompareAndSwap(0, 3)
+		t.CompareAndSwap(0, 2)
+		t.CompareAndSwap(1, 4)
+		t.CompareAndSwap(1, 3)
+		t.CompareAndSwap(1, 2)
+	case 6:
+		t.CompareAndSwap(1, 2)
+		t.CompareAndSwap(0, 2)
+		t.CompareAndSwap(0, 1)
+		t.CompareAndSwap(4, 5)
+		t.CompareAndSwap(3, 5)
+		t.CompareAndSwap(3, 4)
+		t.CompareAndSwap(0, 3)
+		t.CompareAndSwap(1, 4)
+		t.CompareAndSwap(2, 5)
+		t.CompareAndSwap(2, 4)
+		t.CompareAndSwap(1, 3)
+		t.CompareAndSwap(2, 3)
+	case 7:
+		t.CompareAndSwap(1, 2)
+		t.CompareAndSwap(0, 2)
+		t.CompareAndSwap(0, 1)
+		t.CompareAndSwap(3, 4)
+		t.CompareAndSwap(5, 6)
+		t.CompareAndSwap(3, 5)
+		t.CompareAndSwap(4, 6)
+		t.CompareAndSwap(4, 5)
+		t.CompareAndSwap(0, 4)
+		t.CompareAndSwap(0, 3)
+		t.CompareAndSwap(1, 5)
+		t.CompareAndSwap(2, 6)
+		t.CompareAndSwap(2, 5)
+		t.CompareAndSwap(1, 3)
+		t.CompareAndSwap(2, 4)
+		t.CompareAndSwap(2, 3)
+	case 8:
+		t.CompareAndSwap(0, 1)
+		t.CompareAndSwap(2, 3)
+		t.CompareAndSwap(0, 2)
+		t.CompareAndSwap(1, 3)
+		t.CompareAndSwap(1, 2)
+		t.CompareAndSwap(4, 5)
+		t.CompareAndSwap(6, 7)
+		t.CompareAndSwap(4, 6)
+		t.CompareAndSwap(5, 7)
+		t.CompareAndSwap(5, 6)
+		t.CompareAndSwap(0, 4)
+		t.CompareAndSwap(1, 5)
+		t.CompareAndSwap(1, 4)
+		t.CompareAndSwap(2, 6)
+		t.CompareAndSwap(3, 7)
+		t.CompareAndSwap(3, 6)
+		t.CompareAndSwap(2, 4)
+		t.CompareAndSwap(3, 5)
+		t.CompareAndSwap(3, 4)
+	default:
+		sort.Sort(t)
+	}
+}
+
+// Search is the same as sort.Search() but is optimized for our use case.
+func (t tagSet) Search(key string) int {
+	i, j := 0, len(t)
+	for i < j {
+		h := (i + j) / 2
+		if t[h].key < key {
+			i = h + 1
+		} else {
+			j = h
+		}
+	}
+	return i
+}
+
+// Contains returns if the tagSet contains key.
+func (t tagSet) Contains(key string) bool {
+	if len(t) == 0 {
+		return false
+	}
+	i := t.Search(key)
+	return i < len(t) && t[i].key == key
+}
+
+// Inserts tagPair p into tagSet, if a tagPair with the same key exists it is
+// replaced.
+func (t tagSet) Insert(p tagPair) tagSet {
+	i := t.Search(p.key)
+	if i < len(t) && t[i].key == p.key {
+		t[i].value = p.value
+		return t // exists
+	}
+	// append t to the end of the slice
+	if i == len(t) {
+		return append(t, p)
+	}
+	// insert p
+	t = append(t, tagPair{})
+	copy(t[i+1:], t[i:])
+	t[i] = p
+	return t
+}
+
+// mergeTagSets merges s1 into s2 and stores the result in scratch. Both s1 and
+// s2 must be sorted and s2 can be a sub-slice of scratch if it is located at
+// the tail of the slice (this allows us to allocate only one slice).
+func mergeTagSets(s1, s2, scratch tagSet) tagSet {
+	a := scratch
+	i, j, k := 0, 0, 0
+	for ; i < len(s1) && j < len(s2) && k < len(a); k++ {
+		if s1[i].key == s2[j].key {
+			a[k] = s2[j]
+			i++
+			j++
+		} else if s1[i].key < s2[j].key {
+			a[k] = s1[i]
+			i++
+		} else {
+			a[k] = s2[j]
+			j++
+		}
+	}
+	if i < len(s1) {
+		k += copy(a[k:], s1[i:])
+	}
+	if j < len(s2) {
+		k += copy(a[k:], s2[j:])
+	}
+	return a[:k]
+}
+
+func serializeTagSet(name string, set tagSet) string {
+	// NB: the tagSet must being sorted and have clean values
+
+	const prefix = ".__"
+	const sep = "="
+
+	if len(set) == 0 {
+		return name
+	}
+
+	n := (len(prefix)+len(sep))*len(set) + len(name)
+	for _, p := range set {
+		n += len(p.key) + len(p.value)
+	}
+
+	// CEV: this is same as strings.Builder, but is faster and simpler.
+	b := make([]byte, 0, n)
+	b = append(b, name...)
+	for _, p := range set {
+		b = append(b, prefix...)
+		b = append(b, p.key...)
+		b = append(b, sep...)
+		b = append(b, p.value...)
+	}
+	return *(*string)(unsafe.Pointer(&b))
+}
 
 func serializeTags(name string, tags map[string]string) string {
 	const prefix = ".__"
@@ -41,11 +227,11 @@ func serializeTags(name string, tags map[string]string) string {
 			t1 = t0
 			t0 = tagPair{k, replaceChars(v)}
 		}
-		if t0.dimension > t1.dimension {
+		if t0.key > t1.key {
 			t0, t1 = t1, t0
 		}
-		return name + prefix + t0.dimension + sep + t0.value +
-			prefix + t1.dimension + sep + t1.value
+		return name + prefix + t0.key + sep + t0.value +
+			prefix + t1.key + sep + t1.value
 	case 3:
 		var t0, t1, t2 tagPair
 		for k, v := range tags {
@@ -53,18 +239,18 @@ func serializeTags(name string, tags map[string]string) string {
 			t1 = t0
 			t0 = tagPair{k, replaceChars(v)}
 		}
-		if t1.dimension > t2.dimension {
+		if t1.key > t2.key {
 			t1, t2 = t2, t1
 		}
-		if t0.dimension > t2.dimension {
+		if t0.key > t2.key {
 			t0, t2 = t2, t0
 		}
-		if t0.dimension > t1.dimension {
+		if t0.key > t1.key {
 			t0, t1 = t1, t0
 		}
-		return name + prefix + t0.dimension + sep + t0.value +
-			prefix + t1.dimension + sep + t1.value +
-			prefix + t2.dimension + sep + t2.value
+		return name + prefix + t0.key + sep + t0.value +
+			prefix + t1.key + sep + t1.value +
+			prefix + t2.key + sep + t2.value
 	case 4:
 		var t0, t1, t2, t3 tagPair
 		for k, v := range tags {
@@ -73,25 +259,25 @@ func serializeTags(name string, tags map[string]string) string {
 			t1 = t0
 			t0 = tagPair{k, replaceChars(v)}
 		}
-		if t0.dimension > t1.dimension {
+		if t0.key > t1.key {
 			t0, t1 = t1, t0
 		}
-		if t2.dimension > t3.dimension {
+		if t2.key > t3.key {
 			t2, t3 = t3, t2
 		}
-		if t0.dimension > t2.dimension {
+		if t0.key > t2.key {
 			t0, t2 = t2, t0
 		}
-		if t1.dimension > t3.dimension {
+		if t1.key > t3.key {
 			t1, t3 = t3, t1
 		}
-		if t1.dimension > t2.dimension {
+		if t1.key > t2.key {
 			t1, t2 = t2, t1
 		}
-		return name + prefix + t0.dimension + sep + t0.value +
-			prefix + t1.dimension + sep + t1.value +
-			prefix + t2.dimension + sep + t2.value +
-			prefix + t3.dimension + sep + t3.value
+		return name + prefix + t0.key + sep + t0.value +
+			prefix + t1.key + sep + t1.value +
+			prefix + t2.key + sep + t2.value +
+			prefix + t3.key + sep + t3.value
 	default:
 		// n stores the length of the serialized name + tags
 		n := (len(prefix) + len(sep)) * len(tags)
@@ -101,8 +287,8 @@ func serializeTags(name string, tags map[string]string) string {
 		for k, v := range tags {
 			n += len(k) + len(v)
 			pairs = append(pairs, tagPair{
-				dimension: k,
-				value:     replaceChars(v),
+				key:   k,
+				value: replaceChars(v),
 			})
 		}
 		sort.Sort(pairs)
@@ -112,7 +298,7 @@ func serializeTags(name string, tags map[string]string) string {
 		b = append(b, name...)
 		for _, tag := range pairs {
 			b = append(b, prefix...)
-			b = append(b, tag.dimension...)
+			b = append(b, tag.key...)
 			b = append(b, sep...)
 			b = append(b, tag.value...)
 		}