This commit changes the subScope to keep a slice of tags instead of a
map. This means that we no longer modify or "own" the map we're passed,
which has been a source of bugs recently (that gostats owns any map
passed to it was poorly, if at all, documented).

This change also improves performance by 2x and reduces memory usage and
GC pressure as each subScope will no longer be holding a map (which is
sparse and the GC must scan).

This change also moves us closer to passing tags via a slice (instead of
a map) which is a change that I intend to make part of the V1 release
(assuming I ever get around to that).

```
benchmark                                old ns/op     new ns/op     delta
BenchmarkStore_ScopeWithTags-12          1321          646           -51.10%
BenchmarkStore_ScopeNoTags-12            583           179           -69.30%

benchmark                                old allocs     new allocs     delta
BenchmarkStore_ScopeWithTags-12          4              3              -25.00%
BenchmarkStore_ScopeNoTags-12            4              2              -50.00%

benchmark                                old bytes     new bytes     delta
BenchmarkStore_ScopeWithTags-12          544           512           -5.88%
BenchmarkStore_ScopeNoTags-12            304           112           -63.16%
```