@charlievieth 2020-05-23
> Would be nice to ship this behind a flag, not to say I distrust you but the last compile time force upgrades (goingest, idl) have brought me a lot of pain :(

This PR does not change the public API and does fix a subtle bug resulting from users not realizing that `gostats` **owns** any `map` passed to it. That said, you're free to use an older version of `gostats` until this change has been battle tested (though it does significantly increase the test coverage of critical paths (and adds some limited fuzzing). Also, this project is OSS so we do not and will not provide runtime flags to toggle internal behavior (we use releases for that 😉).